Extract agent streaming helpers
