+++ docs/DECISIONS.md
@@ -215,3 +215,4 @@
 - Workspace sheet presentation lives in `QuillCodeWorkspaceSheetsModifier`, not the workspace shell. `WorkspaceSwiftUIView.swift` owns the presentation state bindings and routes typed callbacks; the sheet presenter owns settings, search, keyboard shortcuts, command palette, worktree create/remove, and thread/project rename sheet composition plus their save/cancel dismissal wiring.
 - Tool-result final-answer copy lives in `AgentFinalAnswerBuilder`, not in `AgentRunner`. `AgentRunner.finalAnswer(...)` remains the compatibility entry point, but shell/browser/file/patch/MCP/Computer Use response wording belongs in the focused builder with direct tests and parity gates.
 - Deterministic mock model behavior lives in `MockLLMClient.swift`, not in `Agent.swift`. The mock client can keep local smoke-test and harness heuristics, but the main agent file should stay focused on agent contracts, streaming preview helpers, execution orchestration, and run result types.
+- Streaming action collection and partial assistant-preview parsing live in `AgentActionStreaming.swift`, not in `Agent.swift`. `AgentRunner` and `TrustedRouterLLMClient` should delegate to `AgentActionStreamCollector`, and draft assistant text should go through `AgentActionStreamPreview` so incomplete streaming JSON is handled in one tested helper.
